+++ src/components/FunctionComponents/NetworkGraph/NetworkDiagramSlaveVoyages.tsx
@@ -12,7 +12,7 @@ import { setIsModalCard, setNodeClass } from '@/redux/getCardFlatObjectSlice';
 import { ENSLAVEMENTNODE } from '@/share/CONST_DATA';
 import LOADINGLOGO from '@/assets/sv-logo_v2_notext.svg';
 
-export const NetworkDiagramPeople = ({
+export const NetworkDiagramSlaveVoyages = ({
   widthPercentage = 75,
   heigthPercentage = 65,
 }) => {
