+++ src/components/canscanding/MenuListDropdown.tsx
@@ -28,13 +28,13 @@ import '@/style/homepage.scss';
 
 import { useState, MouseEvent } from 'react';
 import { PaperDraggable } from './PaperDraggable';
-import { setIsChange, setKeyValue } from '@/redux/rangeSliderSlice';
+import { setIsChange, setKeyValue } from '@/redux/getRangeSliderSlice';
 import { setIsChangeAuto } from '@/redux/getAutoCompleteSlice';
 import { setIsOpenDialog } from '@/redux/getScrollPageSlice';
 import { ArrowDropDown } from '@mui/icons-material';
-import AutocompleteBox from '../Voyages/Results/AutocompletedBox';
-import RangeSlider from '../Voyages/Results/RangeSlider';
-import GeoTreeSelected from '../FunctionComponents/GeoTreeSelected';
+import AutocompleteBox from '../FunctionComponents/AutocompletedBox';
+import RangeSlider from '../FunctionComponents/RangeSlider';
+import GeoTreeSelected from '../FunctionComponents/GeoTreeSelect/GeoTreeSelected';
 import { resetAll } from '@/redux/resetAllSlice';
 
 export const MenuListDropdown = () => {
