+++ src/redux/getRangeSliderSlice.ts
