+++ src/utils/functions/formatCount.ts
