+++ src/utils/functions/formatType.ts
