+++ src/utils/functions/pageVariantsFromTop.ts
@@ -1,0 +1,9 @@
+
+export const pageVariantsFromTop = {
+    initial: { opacity: 0, y: -1000 },
+    animate: { opacity: 1, y: 0 },
+};
+export const pageVariantsFromBottom = {
+    initial: { opacity: -1000, y: 0 },
+    animate: { opacity: 0, y: 1 },
+};
